requirements.txt
pandas==1.0.5
streamlit==1.11.1
streamlit_toggle_switch==1.0.2
numpy
seaborn
matplotlib.pyplot
pygwalker
sklearn